sdm660-common: overlay: frameworks: Move onDrawn callback earlier
The new remote animation flow is waiting for the onDrawn callback to
be issues before triggering the exit animation. This means for
wakeAndUnlock flows, onDrawn cannot wait for the exit animation to
begin. Remove all code that put this callback into a pending state.

Also remove BiometricUnlockController delay. This causes a poor
animation experience when unlocking with devices (sunfish) that delay
the unlock.

Fixes: 220340156
Test: atest KeyguardViewMediatorTest && Use rear FPS devices with AoD
on/off

Change-Id: I44db91715b60587aaa594aff5052b22d01c58bf4
Signed-off-by: Pascoato <viniciuspascoato@gmail.com>
Signed-off-by: dotkit <ewprjkt@proton.me>